+++ CHANGELOG.md
@@ -28,8 +28,9 @@ and this project adheres to [Semantic Versioning].
   - added `__mul__` and `__rmul__` method
   - added `conjugate` method
   - added `inverse` method
-  - added `norm` method
+  - added `norm` attribute
   - added `__truediv__` and `__rtruediv__` method
+  - added `trace` attribute
 
 ## [0.0.1] - 2024-06-20